Confirm modal window popping up when user deletes a task

+++ frontend/src/Components/Main/TasksToRender/ItemTask/ItemTask.tsx
@@ -65,6 +65,8 @@ export default function ItemTask(props: Props) {
   }
 
   function deleteTask() {
+    if (!window.confirm('Do you really wanna delete this task?')) return;
+
     toggleLoader(true);
     fetch(`/api/tasks/${props.task.id}`, {
       method: 'DELETE'
